requirements.txt
numpy>=1.23
pandas>=1.5
scikit-learn>=1.2
matplotlib>=3.7
joblib>=1.3
pdfplumber>=0.10
pytesseract>=0.3.10
Pillow>=10.0
xgboost>=1.7
streamlit>=1.28
imbalanced-learn>=0.11
seaborn>=0.12
tensorflow-cpu>=2.13
pytest>=7.0
openai>=1.0.0
python-dotenv>=1.0.0
reportlab>=4.0.0
